% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_lista_de_pessoas_contribuidoras.R
\name{gerar_lista_contrib}
\alias{gerar_lista_contrib}
\title{Gera lista de pessoas contribuidoras no repositório}
\usage{
gerar_lista_contrib()
}
\value{

}
\description{
Gera lista de pessoas contribuidoras no repositório
}
